// src/content/wcag.gen.ts
export const principles = [
  {
    "num": "1",
    "name": "Postrzegalność"
  },
  {
    "num": "2",
    "name": "Funkcjonalność"
  },
  {
    "num": "3",
    "name": "Zrozumiałość"
  },
  {
    "num": "4",
    "name": "Solidność"
  }
] as const;

export const guidelines = [
  {
    "num": "1.1",
    "name": "Alternatywa tekstowa",
    "principle": "1"
  },
  {
    "num": "1.2",
    "name": "Multimedia",
    "principle": "1"
  },
  {
    "num": "1.3",
    "name": "Możliwość adaptacji",
    "principle": "1"
  },
  {
    "num": "1.4",
    "name": "Rozróżnialność",
    "principle": "1"
  },
  {
    "num": "2.1",
    "name": "Dostępność z klawiatury",
    "principle": "2"
  },
  {
    "num": "2.2",
    "name": "Wystarczający czas",
    "principle": "2"
  },
  {
    "num": "2.3",
    "name": "Ataki padaczki",
    "principle": "2"
  },
  {
    "num": "2.4",
    "name": "Możliwość nawigacji",
    "principle": "2"
  },
  {
    "num": "2.5",
    "name": "Metody obsługi",
    "principle": "2"
  },
  {
    "num": "3.1",
    "name": "Możliwość odczytania",
    "principle": "3"
  },
  {
    "num": "3.2",
    "name": "Przewidywalność",
    "principle": "3"
  },
  {
    "num": "3.3",
    "name": "Pomoc przy wprowadzaniu informacji",
    "principle": "3"
  },
  {
    "num": "4.1",
    "name": "Kompatybilność",
    "principle": "4"
  }
] as const;

export const criteria = [
  {
    "id": "1.1.1",
    "w3cId": "non-text-content",
    "name": "Treść nietekstowa",
    "nameSource": "w3c",
    "level": "A",
    "versions": [
      "2.0",
      "2.1",
      "2.2"
    ],
    "guideline": "1.1",
    "principle": "1"
  },
// ...
] as const;